ajout du fichier cini-insights/explian_file.txt
