--- src/Mlflow/mlflow_predict.yaml
@@ -1,17 +1,0 @@
-apiVersion: argoproj.io/v1alpha1
-kind: Workflow
-metadata:
-  generateName: my-workflow-
-spec:
-  entrypoint: my-entrypoint
-  templates:
-  - container:
-      args:
-      - --run_id=e3bad03bcc63472d8ff9ff143af87c4e
-      - --link=https://minio.lab.sspcloud.fr/alimane/diffusion/TMDB_movies.parquet
-      - --model_name=BEST_MODEL
-      - --model_version=1
-      command:
-      - my-command
-      image: docker-image
-    name: my-entrypoint
