--- src/models/my_docfile
@@ -1,11 +1,0 @@
-FROM ubuntu:22.04
-WORKDIR ${HOME}/cine_insights/src/models
-# Install Python
-RUN apt-get -y update && \
-    apt-get install -y python3-pip
-# Install project dependencies
-COPY my_requiremnet.txt .
-RUN pip install -r my_requiremnet.txt
-COPY script.py .
-COPY src ./src
-CMD ["python3", "script.py"]
